fix roundoff in getParallelCrossing
it turns out that sometimes when a line onto a geoline got cutToSize, the crossing coordinates wouldn't be *quite* equal to the original endpoint coordinates, which made it fail to ignore the crossing like it should.  if statement shave been inserted to avoid such roundoff issues in such circumstances, returning instead the exact anser as they should.