simplify applyProjectionToPath
this method was kind of a mess.  there were a bunch of variables that weren't being used, there was this whole pointless queueing system… and the end result was that it expected there to be interruptions in *movetos* that crossed the antimeridian and got upset when they didn't‽  I'm shockd this wasn't more of an issue, tho of course it was apparently a substantial issue judging by how often that one error message came up with toroidal planets.  oh yeah, and the error message is better now.

also, tests.