move some code into MapProjection costructor
this fixes a bug where conic projections over the whole world were broken because even tho it was calling cutToSize on the region of interest before chosing the map bounds, it was'nt doing that when choosing the standard parallel, so the standard parallel was ending up as nan.  by moving this stuff into the Chart constructor we ensure it uses the same cropped region for both.