always provide at least one continent
previusly it always gave the user at least one country to plot, even if it was below the normal size threshold.  instead of doing that, it should always show at least one continent.  so if everything is tiny, the option it gives the user will likely be a little bit bigger.