avoid edge case on the domain edge
there was a really nasty bug that I didn't even really bother to fix.  apparently getEdgeCrossing() can fail to behave consistently when segments of the polygon move along the boundary of the periodic domain ad point is also on the same boundary of the periodic domain, and this specificly happens often when you're testing for the inside outness of a cropped shape.  *but* the problem stems from it trying to do periodicity in such a way that puts two edges essentially on top of each other, and you can avoid it by just turning off periodicity.  because you don't need it for that particular check, as it turns out.  so I did that and it works.  yay.