fix more roundoff issues for discs
one of the tests started failing when I noticed that it was messed up and tried to make it test more effectively.  refactoring the interpolation function ever so slitely caused the problem to go away, tho to be onest I'm not positive I've fully beat this; it's possible I just caused it to not have roundoff error in this one particular case.  whatever, it would take way too much work to track this down and establish it's *not* a problem, so I'll just wait and hope.