make map shape a separate option
I haven't tested this at all!  but if I did this correctly, there should
now be a new "Border shape" option that allows the user to choose
whether they want rectangular map bounds or not, independent of which
projection they choose.  this is long overdue in my opinion, so I'm
happy to finally be implementing it.  it's a shame there's no typescript
on this computer with which to test it, but I really don't want to
install Node.