better central meridians
now it does a simple average over all tiles to choose the central meridian and standard parallel rather than paying excessive attention to poles small islands off to the side.  it still needs a bit more tweaking, but this is a big improvement over what I had, I think, so I'm pushing it.