fix tech tree chronology
bro, I don't know what I was smoking when I made this list, but it was messed up.  I gess at the time I just looked up "when was the first _" on Wikipedia and used that date for each technology, but it resulted in some total nonsense.  the transition away from bronze happening before swords are even invented?
 the transition away from filament lightbulbs happening in the 70s?  car-centric societies in 1769???  the one weird one I can't figure out at all is why I put the invention of nuclear weapons as 1958… it's not the date of the invention of the hydrogen bomb, nor the first soviet bomb... I can't find the significance of this year.  I know the year in which nukes were invented; did I not know it when I wrote this???