fix some tests
apparently I haven't tested the main branch in months.  whatever, it was just a couple of minor issues.  ideally I should examine every process and demonstrate that it cannot possibly change the number of stressd syllables in a word, but that would be annoying to do.  maybe if this turns out to be a recurring issue.