fix toroid distances
the distance() function for toroids has been pretty messed up for a while now.  I'm onestly flabberghasted that I thaut that was a reasonable way to use modulo.  anyway, the new algorithm is probably more ocmplex than it really needs to be, but it doesn't seem to create a noticeable slowdown so it's probably fine.

I'm not sure if I've fully fixd the issue where coastlines are often weerdly alined to the niner equator.  it's definitely lessend but it seems like it's still happening.  but _maybe_ it's just my imagination?  idk man.