refactor vowel harmony
I decided to ditch tense/lax harmony because idk if it's real the way I implemented it but it definitely looks ugly.  I replaced it with [eo]/[iu] harmony like they have in Swahili verbs.  and I refactord the whole vowel harmony to be cleaner and make it easier to make these changes.