replace cyrillic ф with greek φ
it took me a long time to realize there were this many different φ characters.  but now that I do, I really aut to be using the gree one.  I mean, I think "phi" when I type "ɸ", but that's technically not even its name.

huh, it's kind of weird how many tries it took me to make it recompile properly after that.