arbitrary central latitudes for toroids
I adjusted MapProjection to center its latitude domain on a given central latitude rather than having it always be set by the surface itself.  this has no effect for spheroids and planes but makes toroidal maps work better for ROIs near the inner equator.

this change by itself wasn't too much refactoring, but it had some knock-on effects.  first, I redefined Surface.φMin and Surface.φMax to be the max _acceptable_ inputs to functions like rz.  this means that toroids now have φMin = -∞ and φMax = ∞.  this also means that map projections need to be able to deal with coordinates outside the range [-π, π], which meant that the exact domain being used needed to be passed to a lot of the lower-level path utilities (this inspired me to create a Domain class that holds more information than the Domain interface previusly did).  this introduced a slew of roundoff error related issues, which requried me to make some of my functions more robust.  for example, allowing linterp to extrapolate.

twoth, since map projections are no longer using Surface.referenceLatitudes and Surface.cumulDistances, I realized they're really not necessary.  I made referenceLatitudes and cumulAreas private (they still get used for populating the surface) and removed cumulDistances entirely.

third, since the map projections now need to set their own φRef anyway rather than using the Surface's, I went ahead and implemented a fancy adaptive-step-size integration routine that automaticly assigns more resolution to parts of the map that are highly nonlinear (such as the poles on a Mercator map).  this ended up making MapProjection.differentiability much more accurate, which is nice because rereading that function delt me some psychic damage.

so when you factor in that routine and the tests I wrote for it, this commit ended up being quite big.  it also no longer passes all of the tests, so I need to refine the conformal conic projection to have the correct scale and implement a new projection for Discs because my current system (have every map projection resolve to a perfect disc for Disc surfaces) no longer seems tenable.  also worth noting that there are some artifacts around the edge of the map now.  so there's still more work to be done.  all in all it almost certainly would have been easier to go with my original idea of not touching longitude and having it shift latitudes by some amount before putting them into the map projection (so as far as the projection cares latitude would still always be in the range [-π, π]).  but this is probably better in the long run because it means you don't have to worry about two different definitions of latitude (relative to the equator and relative to the standard parallel).