improve the scale of conformal conic projection
I added a bit of code to make the scale of the conformal conic projection correct at the standard parallel.  this isn't quite good enuff, unfortunately.  there's an artifact where you integrate the pole that causes the scale at the pole to always be off by a little more than 10%, which makes the scale rong by about 10% if the standard parallel is the pole.  it's a very minor issue all things considerd, but I'm going to see if I can find an alternative parameterization that will make it stop doing that.