always use good map projection for discs
I refactored the map projection descriptions because I think it's cleaner to have the code that interfaces directly with the gui in the HTML file.  but of course the more substantial change here is that now it always uses the orthographic projection for planar geoids.  it even hides the map projection dropdown when you're looking at a disc!