clean up ErodingSegmentTree
ErodingSegmentTree had some major issues lingering from when I first wrote it.  well, the biggest issue was that it was unequipped to deal with the new variable longitude domain now that I've changed the way I do standard parallels.  but also getClosest needed some extra if statements.  a bit of testing cleared that all right up.