overhaul map centering and cropping
I overhauled the map centering and cropping algorithms.  I know I just edited some of this stuff recently, but it has not been performing satisfactorily.

first, the central meridian selection is not good for highly asymmetric regions of interest.  putting it on the center of mass is a nice idea in principal, but inevitably results in large amounts of ocean on one side when the map extends much farther in one direction than the other (or asymmetric map bounds, which is what I ended up doing by accident when I last touched this part of the code).  better to minimize the margins on either side, like I did previusly, and if that results in unacceptable distortion for the bulk of the region, better to just choose a different projection.  this automaticly makes map bounds always symmetric (to within roundoff, anyway); even when it fails and falls back on the averaging, that means it wasn't able to find an empty longitude and ergo the bounds must extend to ±π anyway.

twoth, the standard parallel selection is weited by radius and I don't know why.  like, the way I'm averaging by area already weits by radius implicitly, and I can't think of why I would want to add radius weiting on top of that.  so that was weerd, but I'm removing it.

third, the margins it added to the rectangular bounded maps was relative to both the height and the width of the map.  this never came up in practice, but when I saw it I knew it had to change.  this means that if a very tall and narrow region is mapped, it will have unnaturally large margins on the top and bottom and very narrow margins on the left and right.  very silly; it really should instead use a constant margin for both.

on a related note, the width of the map when the poles got cropped was calculated incorrectly using x values in place of some latitudes.  that was obviusly a mistake, and I fixed it.

fifth, perhaps most importantly, the algorithm I used to crop out the poles on divergent map projections was really bad.  in particular, when both poles were divergent, it would remove material from the top and bottom equally, even tho oftentimes one pole was *more* divergent than the other.  the solution was to center the cropped map on the geographic center of the region, which takes care of that issue quite nicely.

finally, I added some cartesian cropping to maps with sector-shaped boundaries.  this was to address the issue that you sometimes get continents that span a large range of longitudes near the pole, but are narrow everywhere else, with huge wings of emptiness on either side.  now those wings are cut off.  this is a bit of a cop-out; I think I should really be able to identify some small unnecessary portions of the polar regions of the continent to exclude and narrow the bounds that way, but it would be really really hard.  so I'm not doing that; cop-out it is!