make lineArcIntersections() more robust
my map boundaries were getting screwy because the new central meridian system creates roundoff that means the two bottom vertices of a sector aren't always _exactly_ at the same y value, and it turns out the lineArcIntersections() routine as I had ritten it was highly susceptible to roundoff to the extent that it would detect an intersection sometimes when the arc actually stopped just a few epsilon shy of the line.  I changed it to use angles that are less likely to flip due to roundoff issues.  I also changed spliceSegment to be more robust in a similar manner because that was where I originally thaut the problem was, tho it turns out that it wouldn't have caused problems regardless.  but since the new code is more robust and more readable (only ever so slitely slower since it calls arcCenter, but who cares because there aren't that many arcs that need to be spliced), I figured I may as well leave it.  I also edited the docstrings of some of the geometry.ts functions because they were just wrong and created some tests for lineArcIntersections().