make lineArcIntersections() even more robust
I made sure this function passed the new tests I wrote for it but did not check the plotting.ts tests that also depend on it.  turns out my edited version was no longer detecting cases where an endpoint is coincident with the line segment.  this has now been corrected.