make contains() work for toroidal planets
there was a gap in my tests – the thing I did where when the contains algorithm fails it tries to draw a northward line to infinity and then back at a different longitude didn't always work on toroidal planets where latitude is periodic just like longitude is.  actually, I did test this, but it only fails about half the time so it didn't come up.

the way I fixed it was pretty clever if I do say so myself.  if the contains() function draws a north–south line thru the point in question and checks for intersections with the polygon.  however, sometimes the point isn't latitudinally in line with the polygon.  when that happens, you need to find a different path, preferably also based on infinitely long lines, that connects the point to part of the polygon.  previusly I drew a twoth north–south line at a different longitude, reasoning that there's an effectively continuus path north to infinity and then back south from infinity, but in a geographic coordinate system there isn't always an infinity to which to go.  so instead, I pick a point on the line, and draw an _east–west_ line thru that.  this works because if the initial line failed to reach a conclusion, that means the polygon does not intersect that line, and that every point on that line must therefore have the same tenance as the sample point – either the whole line is in or the whole line is out.  ergo you can repeat the test for any point on the line and get a valid anser.  criticly, rotating the line 90° and choosing your new point intentionally allows you to garantee that your new line will intersect with something.

anyway, as part of this, I had to rite a function that would rotate the polygon 90°, and it turns out there were a couple other places in my code where this was useful, so that let me simplify transformOutput() and getGeoEdgeCrossing().