tweak colors further
added a new blue, and also an outline between the land and sea.  also the rivers have better colors now.  and I made the heightmap colormaps a bit easier on the eyes.  the outline of the map is kind of ugly now sometimes, but I need to fix that later anyway.