better expansion rate calculation
I changed it so that the probability a country conquers a tile is based on the conquered tile's area divided by the length of the edge separating them, not the distance between the two tiles' centers.  this didn't have as big of an effect as I'd hoped, but I think it does generally make the countries more regularly shaped and reduces the number of weird necks.  I also had to adjust the area calculation for this because the old version was not accurate.