top counting lost tiles in population totals
I only noticed this issue because the demographics percentages weren't adding up to 100%.  just goes to show that it pays to pull on threads like those.  tho I suppose looking thru my TODO comments would have gotten the same effect.

I replaced the TreeMap data structure with just a Map because I wasn't getting any of the benefits of a Tree from it.  I vaigly remember knowing this was a silly way to implement it at the time but for whatever reason I ignored it except to add that TODO comment.  this implementation makes much more sense and actually tracks this.arableArea correctly.