remove jQuery and streamline input spinners
I went thru and finally fully removed jquery as a dependency.  this turnd out to be a bigger change than I expected.  it was easy enuff to change the parts of my code that used jquery, but it turns out that bootstrap-input-spinner requires jquery.  of course, changing the parts of that code that used over to equivalent vanilla javascript functions was also not a big deal.  the real problem was that I noticed that bootstrap-input-spinner works in a fundamentally weird way that I don't like, so I couldn't help myself from restructuring the whole thing.

you see, bootstrap-input-spinner is made to work as a script that you tack onto an existing HTML document to make the spinners look pretty.  but for varius reasons the fancy spinners with which you end up aren't actually numeric inputs, but text inputs.  the script can't change the existing numeric inputs to text ones without altering the behavior of any scripts that were expecting to interface with those spinners, so it instead creates a new fancy "spinner" (text input) right next to the original spinner, hides the original spinner, and then binds them together so that any changes in one are reflected in the other.

this works, but I don't like it.  for one, I'm not fully convinced that it will work smoothly with screen-readers.  also, while the script does its best, it doesn't cover all bilt-in functionality of numeric inputs, like the user being able to use the up and down arrow keys to increment and decrement the value.  it's also just weird!

so I changed it to a totally different regime where the fancy HTML element, with the auxiliary buttons and such, is bilt in build_html.py, and thus doesn't need to be constructed by javascript after page-load.  this naturally alleviates the need to keep an original element alongside a newly constructed one.  most importantly, the core element is now a numeric input.  this imposes some limitations – I can't have thousands separators anymore, and the CSS to hide the efault arrow buttons is more complicated.  but it also adds arrow key functionality and greatly reduces the amount of necessary auxiliary code.  you no longer need code to bild the element because that's all in the python script now.  you no longer need to link the two elements because there's only one element.  you no longer need to propagate events between them for the same reason. you no longer need to watch for as many events because most of those are handled automaticly.  you no longer need to do as much locale-aware number formatting because it turns out the bilt-in inputs already do that.  really the auxiliary code now only exists to enforce consistent number formatting (which bilt-in number inputs don't do well) and assine functionality to the + and − buttons.

since this put the complexity of built_html.py over what I consider acceptable for the functionless format, I broke it all up into functions.  the only behavioral change is that there's now a {Build} command that can be used to make fancy spinners.

it also now uses a minus sign by default instead of a hyphen-minus by default.