unignore built files
I'm not exactly sure if this is the correct way to do this, but I think it is.  I feel like the purpose of gitignore is to not clutter up your diffs, which isn't going to be an issue on this branch where I just dump compiled releases.  the alternative that I can see is using github actions to have GitHub's server's build my TypeScript and run build_html.py automaticly, but that sounds hard.