clean up labeling code
just a few things that I had tagged and that were easy to fix.  I reduced the number of variables I have to set in that one loop by consolidating them inside the loop instead of outside it, I changed the coordinate system to always be left-handed, and I made it stop making the labels avoid lines between disconnected loops.  the labels actually look a bit better now.  probably something to do with the coordinate system flip?  I suspected that was causing problems somewhere.