add new debug plot
now you can have it plot the calculated topological skeleton instead of the label.  it looks nice!  and now I feel much more confident in my skeleton calculations.  as part of this I changed the return value of chooseLabelLocation to be a general path instead of a single PathSegment, which I think is probably better in general.