even better path simplification
now it can both refine and simplify in the same shape.  this turns out to be essential for labelling shapes where part of it is very rough and part of it is very simple.

also, now you can plot just the part of the medial axis that was chosen as the best candidate.