forbid illegal label positions!
at last, we arrive at one of the underlying problems with my label placement.  the way it weights the curvature of labels actually incentivizes it to find labels that are just slightly too curved to be feasible.  it's apparently been this way since 2021, ever since I randomly changed it from 6*log(bendRatio) to bendRatio/(1 - bendRatio) in the middle of some huge commit that was mostly about changing the shades of green available and changing the spellings of diphthongs, and just didn't notice that it must have immediately made my labels look so much worse.  whatever, it's fixed now.  now all I need to do is fix the font size and offset the paths from the text's centerline to its baseline.