fewer super bent labels
I increased the label selection weit on bentness, decreased the weit on horizontality, and changed the selection rootine so that it finds the best label and _then_ checks the font size.  this prevents it from bending the label more to try to fit it in at the minimum font size.  if the bend ratio outweys the area, as it often does, then it should respect that decision regardless of the minimum font size.  this has the bonus effect of making it a little bit faster.  and the other bonus effect of making it so it's not constantly throwing error messages anymore.