fix country focused maps
this got totally broken when I changed Civ.tileTree to a Map.  while Map.keys() is an iterable, it's an iterable that can only be used once.  it aut to be converted to a set.  also it should have the water removed before it goes into the central meridian chooser.  also the country names weren't using the specified transcription system.