fix some edge cases of contains()
I think these may have been extant for a while now.  the big thing is that a single parallel stretching across the map would not technicly contain points at λMin.  but now they do.  the solution was I put an 'L' there so that it's not an edge case anymore.