prevent Bonne self-intersection
I set it up so it adjusts the bounds and curvature of the Bonne projection to prevent self-intersection.  I fear I've gone too far, tho, as it starts by limiting the longitude, and it turns out with toroids the center usually ends up being put adjacent to one of the limiting parallels meaning it just decreases the meridians all the way, which isn't great.  I think for better results I want to do something a little smarter as far as deciding how much to limit φ versus λ.there are other more complicated things I could do.  I could try to determine if the self-intersection is in the Cartesian map bounds and let it be if it's not.  I could try to move clipped latitudes to the other side so that the whole 2π of latitudes is shown in theory (tho in that case what you would really want is to show each latitude multiple times as necessary (I've never done this in vector graphics.  it sounds really really hard.).