reverse definition of latitude on disc
this is part of a bigger change to use better names for cardinal directions on tidally locked planets.  I'm not doing that yet, tho, because this part is big enuff I want it in a separate commit.

it may seem strange than any changes to the coordinate systems should be necessary to change the labels on the compass rose, but a really interesting conundrum came up.  I've decided that I think maps of tidally locked worlds should be night-up by default, because that's the direction people will generally prefer to face – especially when surveying terrain.  this is the opposite of how it was before, mandating that the insolation function be changed on tidally locked spheres.  and that naturally applies to discs as well as spheres.  in other words, that means latitude should increase away from the sun.  however, the sun logicly must be above the center of the disc planet, and the disc planet layout is based on modern Flat Earth Theory wherein the center is the _north_ pole and thus latitude increases toward the center.  I could have the Flat Earth convention for normal discs and the night-up convention for tidally locked discs, but this would make the terrain generation seem inverted between the two, which I would rather not do (unless I do a lot of inheritance, which I would also rather not do because it would be hard).

in the end, I decided to just define latitude on discs to always increase away from the south pole.  it'll be unintuitive to some, but I imagine most won't even notice.  think of it as an intentional challenge to the traditional north-hemisphere-centric nature of modern Flat Earth Theory.

one interesting consequence of this is that the 3D rendering of the Disc technicly appears upside down, since it shows north as up for all planets except tidally locked spheres.  I could fix this by making the 3D plots south-up by default (kind of like the physics convention for polar angle), but I don't think it really matters.