new cardinal direction names when tidally locked
now instead of north, east, etc., on a tidally locked planet it will use the terms night for away from the sun, day for toward the sun, deasil for right when facing nightward, and widdershins for left when facing dayward.  there are similar terms in the other languages we support.  this is maybe going to be a little confusing, especially in the other languages where I don't have as strong a grasp on how to communicate made-up ideas, but I think it's fun.  someday I'll add labels to the compass rose and that'll also help a lot.