fix direction vectors on discs
the way I calculate north and east at each tile was apparently kind of hacky for discs.  this new way uses surface.tangent and is thus much more robust, and in particular is no longer flipped for discs with the south pole at the center like the old method was.