use layering in the output SVG
this will make it a bit easier to edit these in inkscape, I think.  unfortunately the labels are now completely broken because somehow this restructuring of the SVG makes its width read as zero.  I'm not going to worry about how exactly that happens and instead try to rewrite this so that it doesn't depend on measuring the width of the SVG.  this is only the most recent experience to convince me that doing so is super sketchy.