fix input spinners sometimes reverting their value
there was a weerd issue here where input spinners stored their values in two places – in a javascript variable and in the HTML element – and they would get out of sync whenever I programaticly altered them.  this fixes that issue by removing the variable and relying on the actual number displayd in the element as the sole representation of the state.