remove tooltips
these have been very helpful as a debugging tool, but they feel unprofessional for a final product.  now that the labels are better (not completely fixed, tho; I need to deal with the duplicate candidates, and sometimes the borders end up encompassing the whole screen still) they're not so necessary.