change World.politicalMap to an instance variable
idk why I did it like this; it's so much easier syntacticly to just hav ea field on Tile for the current government.  maybe I was hesitant to have too many transient values on Tile because I wanted it to be more immutable?  but I already have a field for the current culture, and this is really the same thing… tbf it is kind of sketchy that I have to manually clear those variables whenever I initialize a new World, but hey if I should have been doing that before anyway then /shrug