finish up culture generation
I went thru the culture generation and just gave it some general finishing touches.  the main thing is that I added the nomadic/sedentary distinction.  in certian biomes, a culture may develop as nomadic pastoralists, who are pretty much the same as a sedentary culture for most purposes, except that now the "+nomadic" and "+sedentary" requirements on cultural aspects actually work.  also, a people is forcibly sedentized if it's part of a Civ whose ruling culture is sedentary.  there was a bug where whenever a culture got assimilated by its ruling culture, it would sometimes actually become the 100-years-previus version of the ruling culture, which then appeared as a separate ethnic groop.  I beleve this is the main reason there were so many minority groops that were almost identical to the ruling culture, down to the name.  I fixd that, and also moved culture updating to the end of the century rather than the beginning, so that cultures can properly account for the current political situation.  a related issue was that the way I determined if a culture had a nation state – checking if its homeland was the capital, caused issues with this because you could have two or more peoples beleving themselves to be the ruling class.  there was another bug where attribute names were still getting added to the list of culture attributes even tho I wasn't using any of them, so I fixd that.
 I also noticed that the new dry tundra biome was missing from culture.json.  I fixd that by replacing the specific biome tags with generic tags like "cold" and "dry" and "barren".  part of this was renaming PLAINS to GRASSLAND.

as a more minor note, I realized I've been implicitly assuming there are uncivilized hunter-gatherer tribes everywhere on the map that's not part of a Civ, so I shouldn't describe that unclaimed land as "tierra virgen" – "tierra poco poblada" is more precise.