make empires have smoother shapes
now, if a country mostly surrounds a Tile and has the military strength to claim it, it does so instantly. call it the seege rule.  this greatly reduces the number of deserted holes that end up in the middles of big countries, and gives them nicer looking shapes in general.  it also increases the average country size, which I think is welcome because they've been kind of small for a while, even with the removal of nationalism.