remove continental tentacles
this was somehow completely broken since the day I implemented it, and I somehow never noticed.  it was gouging out these wide and deep trenches right next to every passive margin, resulting in these noodly appendages appearing on every continent.  it wasn't super apparent for 20000km-circumference worlds, but was extremely apparent for 40000km-circumference worlds.  this negative sine fixes the issue.