give mountain ranges high priority
it's actually somewhat common for a small peece of continental crust to collide with a larger one and turn into a broad mountain range rite next to the ocean (see the rocky mountains).  but the priority system I'd set up made this extremely rare.  the result was that small continents would get erased completely by ocean even if they were pressing up agenst another continent, which had the dual effect of making the continental map focuses quite confusing sometimes.  this fixes that issue by adding continent collisions to the high-priority fault type list so that small continents are more likely to have some land left over when the ocean comes for them, even if it's just a wide mountain range on the side of another continent.