new continent finding algorithm
rather than relying on the tectonic plates to define the "continents" that set the map scale and centering, I wrote a new algorithm to intelligently break the world into several reasonable sized continents.  it's, uh, not very good.  I might need to replace this with something else because it's really really bad.

I also changed the spinbutton in the terrain tab from number of "continents" to number of "plates" so that it's not confusing when it differs from the actual number of continents.