ensure rivers don't flow along coast
it would occasionally happen sometimes becouse of RIVER_WIDTH and because this priority queue library I'm using is first-in-first-out in the case of ties.  I fixd it by immediately assining all river outflows into the ocean rather than putting them in the same cue as all the river propagation along land.  it actually simplifies the typing a bit.