new continent finding algorithm
now it's based on drainage basins, and it makes an effort to pick continents that are compact rather than randomly sticking regions together and ending up with these bizzarely shaped things.  I ended up using the outline function, so I moved that out of Chart into a lower level file.