fix an edge case
some off-screen elements were occasionally being enlarged to cover everything because this one line was rong.  I can't figure out why this line was this way to begin with; surely the last time I modified this was after I had introduced the concept of a Domain, so why didn't I do the obvius and correct thing and use the domain as the domain?