add HTML download button
in case someone wants to download the HTML.  you now, I find it interesting how rare it is for HTML documents to be treated as actual documents outside of the internet.  it's a fine format for formatted but not print-ready text.  but I suppose it's rare that nonprogrammers want anything that's not print-ready, hence PDF and DOCX, and programmers are so used to editing plaintext code that it's natural they would gravitate to TXT and MD.

I had to remove some of the namespace specifications from some of these new Element constructors because they came out of the iframe totally rong for some reason.  as far as I can tell they weren't necessary.