delete Pandunia translation
now that Risto has put Pandunia on hiatus to write a book, after having just partially ended the previus hiatus he put it on to make panlexia, after having just deleted 90% of the vocabulary with the promise that he would rebild it with the new word derivation system, I think it's safe to say that Pandunia is dead and never going to be complete and stable.  so there's no longer a point in supporting this charade in my codebase.  no one will ever be able to read my wobpage in Pandunia, and a Pandunia translation is an exercise in futility that will only confuse readers.