multithreading!
now the throbber actually throbs because the gui thread continues to update while a webworker takes care of the map generation.

I thaut it would be a simple fix.  but it was actually a humungus pain.  the non-gui-thread stuff has to all be in a separate file, which cannot share memory with the gui thread and must communicate with it via serialized messages, cannot access the DOM, and cannot even access the DOM classes like SVGGElement.  this means that in addition to splitting main.ts in half and totally rethinking the way I store state, I even needed to write a virtual DOM class to replace all of the SVG and HTML bilding.

there are a few outstanding issues.  The error messages are not all propagating correctly
Plotly isn't going right now.  I can't seem to use Plotly from the worker, so I need to figure out how to pass the surface information to the main thread where it can then be renderd.
I can't use a test text element to measure label widths anymore, so I need to find an alternative to that.  maybe query the main thread every time I need to check a length?  sounds dreadfully complicated.  maybe pre-load the length of every letter?