failed attempt at multithreaded text measurement
I need to calculate the width of each label in the map font to label the map, but the way I was previusly doing it – putting it in the DOM and then querying the browser for its actual width – doesn't work now because you can't access the DOM from a web worker.  I'll revert this now since it's a dead end, but it was enuff work that I want it to at least be on the record.

I thaut I could get around that by having the web worker query the main thread for text measurements whenever it needs them, but the system I thaut up for receiving responses to those queries doesn't work.  I want to be able to just send a query and wait for it to complete, but I can't receive it while I'm waiting because web workers are single-threaded.  this would be pretty easy with shared memory, but javascript doesn't allow that.  so anyway, to do this using the browser, I would haff to totally refactor this so that the map generation can terminate, a text measurement can be received, and then the map generation can be resumed with the partially done map, which sounds really really hard and probably quite difficult to read as well.  I think it would be better if I could find a way to calculate text width without the browser, but I have yet to find a satisfactory way to do that.  I'll _probably_ end up making a list of every possible character and it's width, but let's see if I can find something more elegant.  how does Azgaar do it?